+++ API/Controllers/CampusController.cs
@@ -22,7 +22,7 @@ namespace BookingApp.Controllers
             return Ok(await _service.GetSitesByAccount());
         }
         [HttpPost]
-        public async Task<ActionResult> DeleteUploadFile([FromForm] decimal key)
+        public async Task<ActionResult> DeleteUploadFile([FromForm] int key)
         {
             return Ok(await _service.DeleteUploadFile(key));
         }
