+++ API/DTO/Filter/UserFilter.cs
@@ -3,7 +3,7 @@
     public class UserFilter
     {
         public string Department { get; set; }
-        public decimal Role { get; set; }
+        public int Role { get; set; }
         public bool LDapLogin { get; set; }
     }
 }
